refactor(cli): carve skill group into cli/groups/skill.py
Move the skill group, skill_install, and the shared skill-install
verification helpers (_packaged_skill_source, _verify_skill_install_target,
_SKILL_TOOL_DIRS, _SkillInstallHealth, …) out of the monolith. doctor.py
and the host-install path import the shared helpers from here now; __init__
re-exports them for back-compat. No behavior change.
